+++ pkg/config/tlsconfig/config_test.go
@@ -28,7 +28,8 @@ import (
 	"time"
 
 	"github.com/smarty/assertions"
-	. "go.thethings.network/lorawan-stack/v3/pkg/config/tlsconfig"
+	"go.thethings.network/lorawan-stack/v3/pkg/config/tlsconfig"
+	"go.thethings.network/lorawan-stack/v3/pkg/util/test"
 	"go.thethings.network/lorawan-stack/v3/pkg/util/test/assertions/should"
 )
 
@@ -83,8 +84,8 @@ func TestApplyTLSClientConfig(t *testing.T) {
 	t.Parallel()
 	a := assertions.New(t)
 	caCert, _ := genCert()
-	tlsConfig := &tls.Config{}
-	err := (&Client{
+	tlsConfig := &tls.Config{} //nolint:gosec
+	err := (&tlsconfig.Client{
 		FileReader: mockFileReader{
 			"ca.pem": caCert,
 		},
@@ -96,9 +97,10 @@ func TestApplyTLSClientConfig(t *testing.T) {
 	a.So(tlsConfig.InsecureSkipVerify, should.BeTrue)
 
 	t.Run("Empty", func(t *testing.T) {
+		t.Parallel()
 		a := assertions.New(t)
 		tlsConfig := &tls.Config{} //nolint:gosec
-		err := (&Client{}).ApplyTo(tlsConfig)
+		err := (&tlsconfig.Client{}).ApplyTo(tlsConfig)
 		a.So(err, should.BeNil)
 		a.So(tlsConfig.RootCAs, should.BeNil)
 		a.So(tlsConfig.InsecureSkipVerify, should.BeFalse)
@@ -110,7 +112,7 @@ func TestApplyTLSServerAuth(t *testing.T) {
 	a := assertions.New(t)
 	cert, key := genCert()
 	tlsConfig := &tls.Config{} //nolint:gosec
-	err := (&ServerAuth{
+	err := (&tlsconfig.ServerAuth{
 		Source: "file",
 		FileReader: mockFileReader{
 			"cert.pem": cert,
@@ -128,7 +130,7 @@ func TestApplyTLSClientAuth(t *testing.T) {
 	a := assertions.New(t)
 	cert, key := genCert()
 	tlsConfig := &tls.Config{} //nolint:gosec
-	err := (&ClientAuth{
+	err := (&tlsconfig.ClientAuth{
 		Source: "file",
 		FileReader: mockFileReader{
 			"cert.pem": cert,
@@ -139,4 +141,24 @@ func TestApplyTLSClientAuth(t *testing.T) {
 	}).ApplyTo(tlsConfig)
 	a.So(err, should.BeNil)
 	a.So(tlsConfig.GetClientCertificate, should.NotBeNil)
+}
+
+func TestACMEHosts(t *testing.T) {
+	t.Parallel()
+	a, ctx := test.New(t)
+	acmeConfig := &tlsconfig.ACME{
+		Enable:      true,
+		Endpoint:    "https://acme.example.com/directory",
+		Dir:         "testdata",
+		Email:       "test@example.com",
+		Hosts:       []string{"example.com", "*.example.org"},
+		DefaultHost: "example.com",
+	}
+	manager, err := acmeConfig.Initialize()
+	a.So(err, should.BeNil)
+	a.So(manager.HostPolicy(ctx, "example.com"), should.BeNil)
+	a.So(manager.HostPolicy(ctx, "subdomain.example.com"), should.NotBeNil)
+	a.So(manager.HostPolicy(ctx, "example.net"), should.NotBeNil)
+	a.So(manager.HostPolicy(ctx, "example.org"), should.NotBeNil)
+	a.So(manager.HostPolicy(ctx, "subdomain.example.org"), should.BeNil)
 }