+++ CHANGELOG.md
@@ -13,6 +13,8 @@ For details about compatibility between different releases, see the **Commitment
 
 ### Changed
 
+- Support wildcards in the supported hosts for TLS certifictes obtained via ACME (`tls.acme.hosts`).
+
 ### Deprecated
 
 ### Removed